+++ src/dev/entities/movingObjects/MetalBox.java
@@ -7,7 +7,7 @@ import java.awt.*;
 
 public class MetalBox extends Box {
 
-    public MetalBox(Handler handler, float x, float y)
+    public MetalBox(Handler handler, int x, int y)
     {
         super(handler, x, y);
         boxType = 3;
@@ -33,6 +33,6 @@ public class MetalBox extends Box {
     @Override
     public void render(Graphics g){
 
-        g.drawImage(Assets.metalBox, (int) x, (int) y, null);
+        g.drawImage(Assets.metalBox, x, y, null);
     }
 }
